requirements.txt
# Main
Django==5.1.7
django-jazzmin==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.9
pillow==11.1.0
jalali_core==1.0.0
jdatetime==5.2.0
jedi==0.19.2
djangorestframework_simplejwt==5.5.0

# Utilities
asgiref==3.8.1
inflection==0.5.1
packaging==24.2
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
ipython==9.0.2
ipython_pygments_lexers==1.1.1

